+++ src/encoded/static/components/browse/components/index.js
@@ -6,3 +6,4 @@ export { SelectedFilesController } from './SelectedFilesController';
 export { CustomColumnController, CustomColumnSelector } from './CustomColumnController';
 export { SearchResultTable, defaultColumnBlockRenderFxn, extendColumnDefinitions, defaultColumnDefinitionMap, columnsToColumnDefinitions } from './SearchResultTable';
 export { AboveTableControls } from './AboveTableControls';
+export { ExperimentSetDetailPane } from './ExperimentSetDetailPane';
