+++ src/encoded/static/scss/encoded/modules/_tables.scss
@@ -404,192 +404,6 @@ table.matrix {
 
 $open-expset-table-side-border-color : #e4e4e4;
 
-.expsets-table {
-
-	width: 99.99%;
-
-	> thead {
-		//border-top: 1px solid #ccc;
-		background-color: #22656d;
-		> tr > th {
-			height: 46px;
-			background-color: transparent;
-			font-weight: 500;
-			font-size: 0.875rem;
-			padding-bottom: 6px;
-			color: #fff;
-
-			.column-sort-icon {
-				color: rgba(255,255,255,0.5);
-				text-decoration: none !important;
-				cursor: pointer;
-				&:hover, &:active, &:focus {
-					color: #fff;
-				}
-				&.active {
-					color: #fff;
-				}
-			}
-		}
-	}
-
-	> tbody.expset-section {
-
-		> tr {
-
-			> td {
-				&.dropdown-button-cell {
-					position: relative;
-
-					.expset-button.icon-container {
-						/*
-						width: 100%;
-						height: 100%;
-						border: none;
-						border-radius: 0;
-						*/
-						width: 80%;
-						height: 70%;
-						border: none;
-						border-radius: 15%;
-						background-color: transparent;
-
-						&.btn:focus {
-							outline: none;
-							background-color: transparent;
-						}
-
-						&.btn:hover {
-							background-color: rgba(0,0,0,0.05);
-						}
-
-						> i.icon {
-							/* Re-enable padding if using a Gizmo or BlackTie Icon
-							padding-top: 3px;
-							padding-left: 1px;
-							*/
-							display: inline-block;
-							vertical-align: middle;
-							font-size: 12px;
-							line-height: 20px;
-						}
-					}
-				}
-
-				&:last-child {
-					border-right: none;
-				}
-
-				&:first-child {
-					border-left: none;
-				}
-
-			}
-
-			&.expset-table-row {
-				box-shadow: 0 -4px 5px -3px rgba(0,0,0,0);
-				transition : box-shadow .35s;
-				> td {
-					height: 46px;
-					padding-top: 4px;
-					padding-bottom: 4px;
-					transition : height .5s, border-top-color .35s;
-					&:last-child:after {
-						content: '';
-						height: 100%;
-						position: absolute;
-						right: 0;
-						top: 0;
-						border-right: 1px solid transparent;
-					}
-				}
-			}
-
-		}
-
-		&:first-of-type {
-			> tr > td {
-				border-top-color: transparent;
-			}
-		}
-
-		&:not(.open):hover {
-			> tr > td {
-				/* background-color: rgba(0,0,0,0.05); */
-				background-color: rgba(36, 139, 172, 0.05);
-			}
-		}
-
-		> tr.expset-addinfo-row {
-			> td {
-				padding: 0;
-				border-top: 1px solid #f4f4f4;
-			}
-		}
-
-		&.open {
-			> tr.expset-table-row {
-				background-color: #fff;
-				box-shadow: 0 -3px 5px -3px rgba(0,0,0,0.33);
-				td {
-					//height: 54px;
-					border-top-color: #808080;
-					position: relative;
-
-					&:not(.mono-text):not(.accession){
-						font-weight: 500;
-					}
-
-					&.dropdown-button-cell {
-						//border-bottom-color: #fff;
-						//background-color: #fff;
-						//border-right-color: rgba(0,0,0,0.075);
-						/*
-						&:before {
-							content: '';
-							height: 100%;
-							position: absolute;
-							left: 0;
-							top: 0;
-							border-right: 1px solid $open-expset-table-side-border-color;
-						}
-						*/
-					}
-					&.dropdown-button-cell:after {
-						content: "";
-						position: absolute;
-						background: #fff;
-						left: 0;
-						right: 0;
-						height: 20px;
-						bottom: -20px;
-					}
-					&:last-child:after {
-						content: '';
-						height: 100%;
-						position: absolute;
-						right: 0;
-						top: 0;
-						border-right: 1px solid $open-expset-table-side-border-color;
-					}
-				}
-			}
-
-			+ :not(.open) > tr > td.dropdown-button-cell {
-				border-top: 0px solid transparent;
-			}
-		}
-		&.closed:last-child > tr > td {
-			border-bottom: 1px solid #ddd;
-		}
-	}
-
-	.expset-table-cell {
-		padding: 7px 5px;
-	}
-
-}
-
 .expset-hits {
 	padding: 0px 10px 0px 10px;
 }
@@ -708,13 +522,10 @@ $open-expset-table-side-border-color : #e4e4e4;
 
 		.expset-addinfo-key {
 			font-weight: 500;
-			padding-right: 10px;
-			width: 100px;
-			display: inline-block;
 		}
 
 		.expset-addinfo-val {
-			padding-right: 30px;
+			padding-right: 25px;
 		}
 	}
